requirements.txt
django==4.2.4
django-environ==0.9.0
django-cors-headers==4.2.0
djangorestframework==3.14.0